+++ CHANGELOG.md
@@ -9,6 +9,9 @@ v2.6.5 (FUTURE)
 * [#3464](https://github.com/netbox-community/netbox/issues/3464) -  Fix foreground text color on color picker fields
 * [#3519](https://github.com/netbox-community/netbox/issues/3519) -  Prevent cables from being terminated to virtual/wireless interfaces via API
 * [#3521](https://github.com/netbox-community/netbox/issues/3521) -  Fix error in `parseURL` related to variables in API URL
+* [#3531](https://github.com/netbox-community/netbox/issues/3531) -  Fixed rack role foreground color
+* [#3532](https://github.com/netbox-community/netbox/issues/3532) -  Added "device" to graph documentation under "type" field
+* [#3534](https://github.com/netbox-community/netbox/issues/3534) -  Added blank option for untagged vlans
 
 v2.6.4 (2019-09-19)